enabled show legend switch
