+++ src/vis/scatter/ScatterVisNew.tsx
@@ -12,11 +12,10 @@ import { VIS_NEUTRAL_COLOR, VIS_TRACES_COLOR } from '../general/constants';
 import { EColumnTypes, EScatterSelectSettings, ICommonVisProps } from '../interfaces';
 import { BrushOptionButtons } from '../sidebar/BrushOptionButtons';
 import { ELabelingOptions, ERegressionLineType, IInternalScatterConfig, IRegressionResult } from './interfaces';
-import { fetchColumnData, regressionToAnnotation } from './utilsNew';
+import { defaultRegressionLineStyle, fetchColumnData, regressionToAnnotation } from './utils';
 import { getLabelOrUnknown } from '../general/utils';
 import { columnNameWithDescription, truncateText } from '../general/layoutUtils';
 import { fitRegressionLine } from './Regression';
-import { defaultRegressionLineStyle } from './utils';
 import { useDataPreparation } from './useDataPreparation';
 import { InvalidCols } from '../general/InvalidCols';
 import { i18n } from '../../i18n/I18nextManager';
